Change return type (int -> float) on Axis min & max
Annotations in VDS can be non-integers and the corresponding methods in
OpenVDS return float. Returning / implicitly casting to int here is
incorrect.

+++ internal/vds/axis.cpp
@@ -13,11 +13,11 @@ Axis::Axis(
     m_axis_descriptor(layout->GetAxisDescriptor(dimension))
 {}
 
-int Axis::min() const noexcept(true) {
+float Axis::min() const noexcept(true) {
     return this->m_axis_descriptor.GetCoordinateMin();
 }
 
-int Axis::max() const noexcept(true) {
+float Axis::max() const noexcept(true) {
     return this->m_axis_descriptor.GetCoordinateMax();
 }
